+++ changelog.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased] - SC
 ### Added
+- Views controller
+- by_priority view
+- by_category view
+
+## [Unreleased] - SC
+### Added
 - changelog.md
 - Tasks model